+++ GameEngine/Enemy.cpp
@@ -33,7 +33,7 @@ bool Enemy::IsVisible( float visibleAngle, float range)
 	toPlayer = XMVector3Normalize(enemyParameter_.toPlayerVec);							//³‹K‰»
 
 	XMVECTOR dot = XMVector3Dot(GetFrontVec(), toPlayer);								//“àÏ‚ğŒvZ
-	float angle = acos(min(XMVectorGetX(dot), 1));										//Šp“xŒvZ(1ˆÈã‚É‚È‚ç‚È‚¢‚æ‚¤minŠÖ”‚Â‚¯‚½)
+	float angle = acosf(min(XMVectorGetX(dot), 1));										//Šp“xŒvZ(1ˆÈã‚É‚È‚ç‚È‚¢‚æ‚¤minŠÖ”‚Â‚¯‚½)
 	if (toPlayerRange <= 2 * range)
 	{
 		enemyParameter_.pPlayer->AddTargetList(this);
