+++ GameEngine/Engine/Collider/Collider.cpp
@@ -87,9 +87,9 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 
 	
 	
-	double rA = XMVectorGetX(XMVector3Length(obbA->OBB_X));								    //obbAの中心点からの長さ
-	double rB = obbA->prjLine(&obbA->OBB_X, &obbB->OBB_X, &obbB->OBB_Y, &obbB->OBB_Z);		//obbBの中心点からの長さ
-	float  L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbA->OBB_X))));//中心点間の長さ
+	float rA = XMVectorGetX(XMVector3Length(obbA->OBB_X));								    //obbAの中心点からの長さ
+	float rB = obbA->prjLine(&obbA->OBB_X, &obbB->OBB_X, &obbB->OBB_Y, &obbB->OBB_Z);		//obbBの中心点からの長さ
+	float  L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbA->OBB_X))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -97,7 +97,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 
 	rA = XMVectorGetX(XMVector3Length(obbA->OBB_Y));								    //obbAの中心点からの長さ
 	rB = obbA->prjLine(&obbA->OBB_Y, &obbB->OBB_X, &obbB->OBB_Y, &obbB->OBB_Z);			//obbBの中心点からの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbA->OBB_Y))));//中心点間の長さ
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbA->OBB_Y))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -105,7 +105,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 
 	rA = XMVectorGetX(XMVector3Length(obbA->OBB_Z));								    //obbAの中心点からの長さ
 	rB = obbA->prjLine(&obbA->OBB_Z, &obbB->OBB_X, &obbB->OBB_Y, &obbB->OBB_Z);			//obbBの中心点からの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbA->OBB_Z))));//中心点間の長さ
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbA->OBB_Z))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -113,7 +113,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 
 	rA = XMVectorGetX(XMVector3Length(obbB->OBB_X));								    //obbAの中心点からの長さ
 	rB = obbA->prjLine(&obbB->OBB_X, &obbA->OBB_X, &obbA->OBB_Y, &obbA->OBB_Z);			//obbBの中心点からの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbB->OBB_X))));//中心点間の長さ
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbB->OBB_X))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -121,7 +121,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 
 	rA = XMVectorGetX(XMVector3Length(obbB->OBB_Y));								    //obbAの中心点からの長さ
 	rB = obbA->prjLine(&obbB->OBB_Y, &obbA->OBB_X, &obbA->OBB_Y, &obbA->OBB_Z);			//obbBの中心点からの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbB->OBB_Y))));//中心点間の長さ
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbB->OBB_Y))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -129,7 +129,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 
 	rA = XMVectorGetX(XMVector3Length(obbB->OBB_Z));								    //obbAの中心点からの長さ
 	rB = obbA->prjLine(&obbB->OBB_Z, &obbA->OBB_X, &obbA->OBB_Y, &obbA->OBB_Z);			//obbBの中心点からの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbB->OBB_Z))));//中心点間の長さ
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obbB->OBB_Z))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -142,7 +142,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nX, obbB->nX));
 	rA = obbA->prjLine(&cross, &obbA->OBB_Y, &obbA->OBB_Z);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_Y, &obbB->OBB_Z);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -152,7 +152,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nX, obbB->nY));
 	rA = obbA->prjLine(&cross, &obbA->OBB_Y, &obbA->OBB_Z);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_X, &obbB->OBB_Z);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -162,7 +162,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nX, obbB->nZ));
 	rA = obbA->prjLine(&cross, &obbA->OBB_Y, &obbA->OBB_Z);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_X, &obbB->OBB_Y);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -172,7 +172,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nY, obbB->nX));
 	rA = obbA->prjLine(&cross, &obbA->OBB_X, &obbA->OBB_Z);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_Y, &obbB->OBB_Z);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -182,7 +182,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nY, obbB->nY));
 	rA = obbA->prjLine(&cross, &obbA->OBB_X, &obbA->OBB_Z);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_X, &obbB->OBB_Z);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -192,7 +192,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nY, obbB->nZ));
 	rA = obbA->prjLine(&cross, &obbA->OBB_X, &obbA->OBB_Z);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_X, &obbB->OBB_Y);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -202,7 +202,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nZ, obbB->nX));
 	rA = obbA->prjLine(&cross, &obbA->OBB_X, &obbA->OBB_Y);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_Y, &obbB->OBB_Z);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -212,7 +212,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nZ, obbB->nY));
 	rA = obbA->prjLine(&cross, &obbA->OBB_X, &obbA->OBB_Y);				//obbAの長さ
 	rB = obbA->prjLine(&cross, &obbB->OBB_X, &obbB->OBB_Z);				//obbBの長さ
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));	//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -222,7 +222,7 @@ bool Collider::IsHitOBB_OBB(OBBCollider* obbA, OBBCollider* obbB)
 	cross = XMVector3Normalize(XMVector3Cross(obbA->nZ, obbB->nZ));
 	rA = obbA->prjLine(&cross, &obbA->OBB_X, &obbA->OBB_Y);
 	rB = obbA->prjLine(&cross, &obbB->OBB_X, &obbB->OBB_Y);
-	L  = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));//分離軸に投影した中心点間の距離
+	L  = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, cross)));//分離軸に投影した中心点間の距離
 	if (L > rA + rB)
 	{
 		return false;
@@ -244,7 +244,7 @@ bool Collider::IsHitOBB_Sphere(OBBCollider* obb, SphereCollider* sphere)
 	float L;
 	double rA = XMVectorGetX(XMVector3Length(obb->OBB_X));								    //obbの中心点からの長さ
 	double rB = sphere->size_.x;															//sphereの中心点からの長さ
-	L = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obb->OBB_X))));//中心点間の長さ
+	L = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obb->OBB_X))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -252,7 +252,7 @@ bool Collider::IsHitOBB_Sphere(OBBCollider* obb, SphereCollider* sphere)
 
 	rA = XMVectorGetX(XMVector3Length(obb->OBB_Y));								    //obbAの中心点からの長さ
 	rB = sphere->size_.x;															//obbBの中心点からの長さ
-	L = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obb->OBB_Y))));//中心点間の長さ
+	L = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obb->OBB_Y))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
@@ -260,7 +260,7 @@ bool Collider::IsHitOBB_Sphere(OBBCollider* obb, SphereCollider* sphere)
 
 	rA = XMVectorGetX(XMVector3Length(obb->OBB_Z));								    //obbAの中心点からの長さ
 	rB = sphere->size_.x;															//obbBの中心点からの長さ
-	L = fabs(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obb->OBB_Z))));//中心点間の長さ
+	L = fabsf(XMVectorGetX(XMVector3Dot(betweenCenterPoint, XMVector3Normalize(obb->OBB_Z))));//中心点間の長さ
 	if (L > rA + rB)
 	{
 		return false;
